refactor(toml)!: setup structure for serialize implementation
