+++ facet-toml/tests/deserialize/mod.rs
@@ -1,0 +1,8 @@
+mod basic;
+mod document;
+mod enum_;
+mod list;
+mod map;
+mod option;
+mod scalar;
+mod struct_;
